requirements.txt
scikit-learn==1.1.1
easydict==1.9
flow-vis==0.1
matplotlib==3.5.2
natsort==8.2.0
ninja==1.11.1
numpy==1.24.2
opencv-python==4.6.0.66
Pillow==9.2.0
PyYAML==6.0
tensorboardX==2.5.1
torch-encoding==1.2.1
tqdm==4.64.0
scikit-image==0.20.0
# open3d